+++ include/network.h
@@ -6,7 +6,7 @@
 #include"matrix.h"
 using namespace std;
 
-const float ETA_0=1, ALPHA=1, ETA_MIN=0.000001;
+const double ETA_0=1, ALPHA=1, ETA_MIN=0.000001;
 const int SAMPLE=50;
 const int INTERVAL_A=-4, INTERVAL_B=4;
 
@@ -15,10 +15,10 @@ class labeled_data{
 public: 
     labeled_data(int batch_size, int input_size, int output_size);
     labeled_data(){};
-    void setinput(int position, int coord, float value);
-    float getinput(int position, int coord);
-    void setoutput(int position, int coord, float value);
-    float getoutput(int position, int coord);
+    void setinput(int position, int coord, double value);
+    double getinput(int position, int coord);
+    void setoutput(int position, int coord, double value);
+    double getoutput(int position, int coord);
     matrix getinputmatrix(int position);
     matrix getoutputmatrix(int position);
     int size();
@@ -35,17 +35,17 @@ public:
     network(vector<int> sizes);
     network(){};
     vector<int> sizes();
-    void setweight(int layer, int colmn, int row, float value);
+    void setweight(int layer, int colmn, int row, double value);
     void setweights(int layer, matrix value);
-    void setbias(int layer, int colmn, int row, float value);
+    void setbias(int layer, int colmn, int row, double value);
     void setbiases(int layer, matrix value);
-    float getweight(int layer, int colmn, int row);
+    double getweight(int layer, int colmn, int row);
     matrix getweights(int layer);
-    float getbias(int layer, int colmn, int row);
+    double getbias(int layer, int colmn, int row);
     matrix getbiases(int layer);
     network sum(network B);
-    network by(float scalar);
-    float dot(network);
+    network by(double scalar);
+    double dot(network);
     matrix feedforward(matrix a);
     network backprop(matrix input, matrix output);
     void GD(labeled_data mini_batch);
